+++ core/src/main/java/io/snabble/sdk/assetservice/AssetService.kt
@@ -9,18 +9,16 @@ import android.util.DisplayMetrics
 import com.caverock.androidsvg.SVG
 import io.snabble.sdk.Project
 import io.snabble.sdk.assetservice.assets.data.AssetsRepositoryImpl
-import io.snabble.sdk.assetservice.data.ImageRepositoryImpl
-import io.snabble.sdk.assetservice.assets.data.source.dto.AssetDto
+import io.snabble.sdk.assetservice.image.data.ImageRepositoryImpl
 import io.snabble.sdk.assetservice.assets.data.source.LocalAssetDataSourceImpl
-import io.snabble.sdk.assetservice.data.local.image.LocalDiskDataSourceImpl
-import io.snabble.sdk.assetservice.data.local.image.LocalMemorySourceImpl
+import io.snabble.sdk.assetservice.image.data.local.image.LocalDiskDataSourceImpl
+import io.snabble.sdk.assetservice.image.data.local.image.LocalMemorySourceImpl
 import io.snabble.sdk.assetservice.assets.data.source.RemoteAssetsSourceImpl
 import io.snabble.sdk.assetservice.assets.domain.AssetsRepository
 import io.snabble.sdk.assetservice.assets.domain.model.Asset
-import io.snabble.sdk.assetservice.domain.ImageRepository
-import io.snabble.sdk.assetservice.domain.model.Type
-import io.snabble.sdk.assetservice.domain.model.UiMode
-import io.snabble.sdk.extensions.xx
+import io.snabble.sdk.assetservice.image.domain.ImageRepository
+import io.snabble.sdk.assetservice.image.domain.model.Type
+import io.snabble.sdk.assetservice.image.domain.model.UiMode
 import io.snabble.sdk.utils.Logger
 import java.io.InputStream
 import kotlin.math.roundToInt
