+++ core/src/main/java/io/snabble/sdk/assetservice/assets/domain/AssetsRepository.kt
@@ -1,9 +1,8 @@
 package io.snabble.sdk.assetservice.assets.domain
 
-import io.snabble.sdk.assetservice.assets.data.source.dto.AssetDto
 import io.snabble.sdk.assetservice.assets.domain.model.Asset
-import io.snabble.sdk.assetservice.domain.model.Type
-import io.snabble.sdk.assetservice.domain.model.UiMode
+import io.snabble.sdk.assetservice.image.domain.model.Type
+import io.snabble.sdk.assetservice.image.domain.model.UiMode
 
 interface AssetsRepository {
